@@ -7,4 +7,4 @@ build-backend = "setuptools.build_meta"
 [tool.black]
 
 line-length = 80
-target-version = ["py39"]
+target-version = ["py310"]
