@@ -2,6 +2,7 @@
 Tests for :mod:`ranger-ims-server.ext.json`
 """
 
+from collections.abc import Callable
 from datetime import date as Date
 from datetime import datetime as DateTime
 from datetime import timedelta as TimeDelta
@@ -9,7 +10,7 @@ from datetime import timezone as TimeZone
 from io import BytesIO
 from json import JSONDecodeError
 from textwrap import dedent
-from typing import Any, Callable, cast
+from typing import Any, cast
 
 from hypothesis import given
 from hypothesis.strategies import composite, dates
